convert OrderWithinStage from int to short
We'll never need the full range of integers.
Converting to short will allow us to represent
stages as a range within int if that becomes
more useful down the road.

Also added a few more MiddlewareStep instances
and starting all steps at 0 instead of MinValue.

I think it's easier to reason about.

+++ CommandDotNet.Tests/FeatureTests/CommandInvocatorTest.cs
@@ -128,11 +128,11 @@ namespace CommandDotNet.Tests.FeatureTests
 
             if (postBindValues != null)
             {
-                appRunner.Configure(c => c.UseMiddleware(postBindValues, MiddlewareStages.PostBindValuesPreInvoke, int.MaxValue));
+                appRunner.Configure(c => c.UseMiddleware(postBindValues, MiddlewareStages.PostBindValuesPreInvoke, short.MaxValue));
             }
             if (preBindValues != null)
             {
-                appRunner.Configure(c => c.UseMiddleware(preBindValues, MiddlewareStages.PostParseInputPreBindValues, int.MaxValue));
+                appRunner.Configure(c => c.UseMiddleware(preBindValues, MiddlewareStages.PostParseInputPreBindValues, short.MaxValue));
             }
 
             var args = $"NotifyOwner --Number {carNumber} --owner {ownerName}".SplitArgs();
